[BIR] Add modCount to BirImplChildElementList, don't compress its other fields

+++ compiler/ir/bir/tree/src/org/jetbrains/kotlin/bir/BirImplElementBase.kt
@@ -103,7 +103,7 @@ abstract class BirImplElementBase(elementClass: BirElementClass<*>) : BirElement
                     containingList.parent.throwChildForReplacementNotFound(this)
                 }
 
-                return containingList.id
+                return containingList.id.toInt()
             } else {
                 return parent.replaceChildProperty(this, new)
             }
